feat: Add deletePopo, Exception Class
API, Service 기능 추가
    포포 삭제 추가 (DELETE /popo/:popoId)
    포포 배경 수정 (PATCH /popo/:popoId/background)

예외 처리
    일반적인 서버 상의 예외 들을 Global Exception에 추가
    사용자 정의 예외를 추가

다음 계획
    @Valid 와 사용자 정의 예제를 통한 PopoService 예외처리
    Test 코드 생성, 단위 테스트 진행

+++ popo/src/main/java/com/fortice/popo/domain/popo/dao/OptionDAO.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OptionDAO extends JpaRepository<Option, Long> {
+public interface OptionDAO extends JpaRepository<Option, Integer> {
 }
